Refactored Parser by adding traits
