+++ src/Parser.php
@@ -4,10 +4,18 @@ declare(strict_types=1);
 
 namespace Serhii\GoodbyeHtml;
 
-use Exception;
+use Serhii\GoodbyeHtml\Syntax\ReplacesIf;
+use Serhii\GoodbyeHtml\Syntax\ReplacesIfElse;
+use Serhii\GoodbyeHtml\Syntax\ReplacesTernary;
+use Serhii\GoodbyeHtml\Syntax\ReplacesVariables;
 
 final class Parser
 {
+    use ReplacesIfElse;
+    use ReplacesIf;
+    use ReplacesTernary;
+    use ReplacesVariables;
+
     /**
      * @var string The content that is being parsed.
      */
@@ -46,7 +54,6 @@ final class Parser
         $this->replaceIfElseStatementsFromHtml();
         $this->replaceIfStatementsFromHtml();
         $this->replaceTernaryStatementsFromHtml();
-        $this->replaceTernaryVariablesFromHtml();
         $this->replaceVariablesFromHtml();
 
         return $this->html_content;
@@ -64,85 +71,6 @@ final class Parser
     {
         $this->html_content = str_replace($parsed['raw'], $parsed['replacements'], $this->html_content);
         $this->removeUsedVariables($parsed['var_names']);
-    }
-
-    private function replaceTernaryVariablesFromHtml(): void
-    {
-        preg_match_all(Regex::TERNARY_VARIABLES, $this->html_content, $matches);
-
-        [$raw, $var_names, $true_block,,,,$false_block] = $matches;
-
-        $with_raw = $this->getVarNamesWithRaw($raw, $var_names, $true_block, $false_block);
-
-        $new_replacements = [];
-
-        foreach ($with_raw['replacements'] as $key => $item) {
-            $first_symbol = $item[0] ?? null;
-            $new_replacements[$key] = $first_symbol === '$' ? "{{ $item }}" : trim($item, "'\"");
-        }
-
-        $with_raw['replacements'] = $new_replacements;
-
-        $this->replaceStatements($with_raw);
-    }
-
-    private function replaceIfElseStatementsFromHtml(): void
-    {
-        preg_match_all(Regex::IF_ELSE_STATEMENTS, $this->html_content, $matches);
-
-        [$raw, $var_names, $true_block, $false_block] = $matches;
-
-        $this->replaceStatements($this->getVarNamesWithRaw($raw, $var_names, $true_block, $false_block));
-    }
-
-    private function replaceTernaryStatementsFromHtml(): void
-    {
-        preg_match_all(Regex::TERNARY_STATEMENTS, $this->html_content, $matches);
-
-        [$raw, $var_names,,, $true_block,,, $false_block] = $matches;
-
-        $this->replaceStatements($this->getVarNamesWithRaw($raw, $var_names, $true_block, $false_block));
-    }
-
-    private function replaceIfStatementsFromHtml(): void
-    {
-        preg_match_all(Regex::IF_STATEMENTS, $this->html_content, $matches);
-
-        [$raw, $var_names, $contents] = $matches;
-
-        $replacements = [];
-
-        for ($i = 0; $i < count($raw); $i++) {
-            if ($this->variables[$var_names[$i]]) {
-                $replacements[] = trim($contents[$i]);
-            }
-        }
-
-        $this->replaceStatements(compact('raw', 'replacements', 'var_names'));
-    }
-
-    /**
-     * @throws \Exception
-     */
-    private function replaceVariablesFromHtml(): void
-    {
-        preg_match_all(Regex::VARIABLES, $this->html_content, $matches);
-
-        [$raw, $var_names] = $matches;
-
-        $replacements = [];
-
-        for ($i = 0; $i < count($raw); $i++) {
-            $var_key = $var_names[$i];
-
-            if ( ! in_array($var_key, array_keys($this->variables))) {
-                throw new Exception("Undefined variable \${$var_key}");
-            }
-
-            $replacements[] = $this->variables[$var_key];
-        }
-
-        $this->replaceStatements(compact('raw', 'replacements', 'var_names'));
     }
 
     private function removeUsedVariables(array $used_vars): void
